Add removers for CourseLearningMaterial
Added MeshDescriptorsEntity trait and applied it to Course, Objectives,
Session, SessionLearningMaterial entities.
