+++ src/Ilios/CoreBundle/Entity/CourseLearningMaterialInterface.php
@@ -6,6 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
 use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
+use Ilios\CoreBundle\Traits\MeshDescriptorsEntityInterface;
 
 /**
  * Interface CourseLearningMaterialInterface
@@ -13,7 +14,8 @@ use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
  */
 interface CourseLearningMaterialInterface extends
     IdentifiableEntityInterface,
-    LoggableEntityInterface
+    LoggableEntityInterface,
+    MeshDescriptorsEntityInterface
 {
     /**
      * @param string $notes
@@ -64,19 +66,4 @@ interface CourseLearningMaterialInterface extends
      * @return LearningMaterialInterface
      */
     public function getLearningMaterial();
-
-    /**
-     * @param Collection $meshDescriptors
-     */
-    public function setMeshDescriptors(Collection $meshDescriptors);
-
-    /**
-     * @param MeshDescriptorInterface $meshDescriptor
-     */
-    public function addMeshDescriptor(MeshDescriptorInterface $meshDescriptor);
-
-    /**
-     * @return ArrayCollection|MeshDescriptorInterface[]
-     */
-    public function getMeshDescriptors();
 }
