+++ src/Ilios/CoreBundle/Entity/SessionLearningMaterialInterface.php
@@ -5,6 +5,7 @@ namespace Ilios\CoreBundle\Entity;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
+use Ilios\CoreBundle\Traits\MeshDescriptorsEntityInterface;
 
 /**
  * Interface SessionLearningMaterialInterface
@@ -13,7 +14,8 @@ use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
 interface SessionLearningMaterialInterface extends
     IdentifiableEntityInterface,
     LoggableEntityInterface,
-    SessionStampableInterface
+    SessionStampableInterface,
+    MeshDescriptorsEntityInterface
 {
     /**
      * @param string $notes
@@ -64,19 +66,4 @@ interface SessionLearningMaterialInterface extends
      * @return LearningMaterialInterface
      */
     public function getLearningMaterial();
-
-    /**
-     * @param Collection $meshDescriptors
-     */
-    public function setMeshDescriptors(Collection $meshDescriptors);
-
-    /**
-     * @param MeshDescriptorInterface $meshDescriptor
-     */
-    public function addMeshDescriptor(MeshDescriptorInterface $meshDescriptor);
-
-    /**
-     * @return ArrayCollection|MeshDescriptorInterface[]
-     */
-    public function getMeshDescriptors();
 }
