+++ src/Ilios/CoreBundle/Traits/MeshDescriptorsEntity.php
@@ -1,0 +1,53 @@
+<?php
+
+namespace Ilios\CoreBundle\Traits;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+
+use Ilios\CoreBundle\Entity\MeshDescriptorInterface;
+
+/**
+ * Class MeshDescriptorsEntity
+ * @package Ilios\CoreBundle\Traits
+ */
+trait MeshDescriptorsEntity
+{
+    /**
+     * @param Collection $meshDescriptors
+     */
+    public function setMeshDescriptors(Collection $meshDescriptors)
+    {
+        $this->meshDescriptors = new ArrayCollection();
+
+        foreach ($meshDescriptors as $meshDescriptor) {
+            $this->addMeshDescriptor($meshDescriptor);
+        }
+    }
+
+    /**
+     * @param MeshDescriptorInterface $meshDescriptor
+     */
+    public function addMeshDescriptor(MeshDescriptorInterface $meshDescriptor)
+    {
+        if (!$this->meshDescriptors->contains($meshDescriptor)) {
+            $this->meshDescriptors->add($meshDescriptor);
+        }
+    }
+
+    /**
+     * @param MeshDescriptorInterface $meshDescriptor
+     */
+    public function removeMeshDescriptor(MeshDescriptorInterface $meshDescriptor)
+    {
+        $this->meshDescriptors->removeElement($meshDescriptor);
+    }
+
+    /**
+    * @return MeshDescriptorInterface[]|ArrayCollection
+    */
+    public function getMeshDescriptors()
+    {
+        return $this->meshDescriptors;
+    }
+}
